MongeezDao: add support for MongoDB 2.4
MongoDB used to support treating arbitrary values in an index definition as "include this field in the index".
As of 2.4, it no longer does.  We were using this (albeit unintentionally) in our indexes for changeset executions.
This change makes it so that the indexes can be correctly created on MongoDB 2.4.
It also drops any of the old no-longer-supported indexes so that they stop producing warning messages.

http://docs.mongodb.org/manual/release-notes/2.4-index-types/